kafka-lab/requirements.txt
kafka-python==2.0.2
feedparser==6.0.10
beautifulsoup4==4.12.3
requests==2.31.0
openai>=1.0.0
